- nullterminated range comparison fixes  - made property de-serialization more robust  - greatly optimized property de-serialization

+++ lean/header/lean/xml/xml_file.h
@@ -26,7 +26,7 @@ namespace impl
 	{
 		raw_file file(fileName, file::read);
 		
-		size_t fileSize = file.size();
+		size_t fileSize = static_cast<size_t>(file.size());
 		char *source = document.allocate_string(nullptr, fileSize);
 		
 		file.read(source, fileSize); // TODO: strict (exception on error)
